[SCB-1210] improve the precision of average latency metrics in low tps situation

+++ metrics/metrics-core/src/main/java/org/apache/servicecomb/metrics/core/publish/PublishUtils.java
@@ -34,7 +34,7 @@ public final class PublishUtils {
 
   public static PerfInfo createPerfInfo(MeasurementNode stageNode) {
     PerfInfo perfInfo = new PerfInfo();
-    perfInfo.setTps((int) stageNode.findChild(Statistic.count.name()).summary());
+    perfInfo.setTps(stageNode.findChild(Statistic.count.name()).summary());
     perfInfo.setMsTotalTime(stageNode.findChild(Statistic.totalTime.name()).summary() * 1000);
     // when UT with DefaultRegistry, there is no max value
     MeasurementNode maxNode = stageNode.findChild(Statistic.max.name());
